refactor(anyharness): split checkpoint hook and test suites for repo shape (Lane H)
Repo shape (PROD-SIZE-1) flagged six files this branch grew. The two big
overages split along real seams: the turn-start capture/finalize methods
move to sessions/runtime/checkpoint_hook.rs, the fork-linkage suite moves
to checkpoint_linkage_tests.rs, and the checkpoint retention suite moves
to retention_tests.rs. The small wiring residuals (app/mod.rs +14,
error.rs +7, fork.rs +13, prompt.rs call sites +48, tests.rs assertions
+21) take documented ratchet bumps in lints/anyharness/ratchets.toml in
the entries' existing chronicle style.
